pyproject.toml
[tool.poetry]
name = "jphme-website"
version = "0.1.0"
description = "Jan Philipp Harries personal website with image thumbnail generator"
authors = ["Jan Philipp Harries"]
readme = "README.md"
packages = [{include = "*.py"}]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^10.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate-thumbnails = "generate_thumbnails:main" 